update: 예외 상황 추가

+++ src/party/servises/party.cancel.service.ts
@@ -1,8 +1,7 @@
-import { ForbiddenException, Injectable } from "@nestjs/common";
+import { BadRequestException, ForbiddenException, Injectable, InternalServerErrorException, NotFoundException } from "@nestjs/common";
 import { PartyRepository } from "../repositories/party.repository";
 import { PartyCancelUseCase } from "../usecase/party.cancel.usecase";
 import { Override } from "src/utils/decorators/override.decorator";
-import { todo } from "node:test";
 import { PartyCancelDto } from "../dto/party-cancel.dto";
 
 @Injectable()
@@ -13,14 +12,21 @@ export class PartyCancelService implements PartyCancelUseCase {
     }
 
     @Override()
-    async cancel(id: number, dto: PartyCancelDto) {
-        const currentUser = dto.user;
-        const party = await this.repository.findById(id);
-
-        if (currentUser.studentId !== party.leader.studentId) {
-            throw new ForbiddenException('해당 파티를 삭제할 수 없습니다.');
-        }
-
-        await this.repository.delete(party);
+async cancel(id: number, dto: PartyCancelDto) {
+    const currentUser = dto.user;
+    if (!currentUser) {
+        throw new BadRequestException('유저를 식별하지 못했습니다.');
     }
+
+    const party = await this.repository.findById(id);
+    if (!party) {
+        throw new NotFoundException(`파티를 찾지 못했습니다.`);
+    }
+
+    if (currentUser.studentId !== party.leader.studentId) {
+        throw new ForbiddenException('해당 파티를 삭제할 수 없습니다.');
+    }
+
+    await this.repository.delete(party);
+}
 }
